requirement.txt
asgiref==3.2.10
autopep8==1.4.4
Babel==2.9.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
confusable-homoglyphs==3.2.0
defusedxml==0.6.0
dj-static==0.0.6
Django==3.1.5
django-allauth==0.39.1
django-cors-headers==3.7.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-registration==3.1.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.10
mysqlclient==2.0.3
oauthlib==3.0.1
pep8==1.7.1
Pillow==6.2.2
pkg-resources==0.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.5.0
PyJWT==2.0.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2020.1
razorpay==1.2.0
requests==2.24.0
requests-oauthlib==1.2.0
six==1.15.0
soupsieve==2.2
sqlparse==0.3.1
static3==0.7.0
stripe==2.27.0
tkcalendar==1.6.1
urllib3==1.25.10
waitress==2.0.0
whitenoise==5.2.0
